+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/Certificate.cs
@@ -6,35 +6,29 @@ using System.Text.Json;
 
 namespace Azure.Security.KeyVault.Certificates
 {
-    public class CertificateWithPolicy : Certificate
-    {
-        public CertificatePolicy Policy { get; private set; }
-
-        private const string PolicyPropertyName = "policy";
-
-        internal override void ReadProperty(JsonProperty prop)
-        {
-            switch (prop.Name)
-            {
-                case PolicyPropertyName:
-                    Policy = new CertificatePolicy();
-                    ((IJsonDeserializable)Policy).ReadProperties(prop.Value);
-                    break;
-                default:
-                    base.ReadProperty(prop);
-                    break;
-            }
-        }
-    }
-
+    /// <summary>
+    /// An Azure Key Vault certificate
+    /// </summary>
     public class Certificate : CertificateBase
     {
+        /// <summary>
+        /// The id of the key vault Key backing the certifcate
+        /// </summary>
         public string KeyId { get; private set; }
 
+        /// <summary>
+        /// The id of the key vault Secret which contains the PEM of PFX formatted content of the certficate and it's private key
+        /// </summary>
         public string SecretId { get; private set; }
 
+        /// <summary>
+        /// The content type of the key vault Secret corresponding to the certificate
+        /// </summary>
         public CertificateContentType ContentType { get; private set; }
 
+        /// <summary>
+        /// The CER formatted public X509 certificate
+        /// </summary>
         public byte[] CER { get; private set; }
 
         private const string KeyIdPropertyName = "kid";
