+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/Contact.cs
@@ -6,11 +6,25 @@ using System.Text.Json;
 
 namespace Azure.Security.KeyVault.Certificates
 {
+    /// <summary>
+    /// A contact for certificate management issues for a key vault
+    /// </summary>
     public class Contact : IJsonDeserializable, IJsonSerializable
     {
+        /// <summary>
+        /// Email address of the contact
+        /// </summary>
         public string Email { get; set; }
+        /// <summary>
+        /// Name of the contact
+        /// </summary>
         public string Name { get; set; }
+
+        /// <summary>
+        /// Phone number of the contact
+        /// </summary>
         public string Phone { get; set; }
+
         private const string NamePropertyName = "name";
         private const string EmailPropertyName = "email";
         private const string PhonePropertyName = "phone";
