+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/KeyCurveName.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT License. See License.txt in the project root for
 // license information.
 
-using System;
-
 namespace Azure.Security.KeyVault.Certificates
 {
     /// <summary>
