+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/RsaKeyOptions.cs
@@ -6,13 +6,23 @@ using System.Text.Json;
 
 namespace Azure.Security.KeyVault.Certificates
 {
+    /// <summary>
+    /// Propeties of an RSA key backing a certificate
+    /// </summary>
     public class RsaKeyOptions : KeyOptions
     {
         private const string KeySizePropertyName = "key_size";
         private static readonly JsonEncodedText KeySizePropertyNameBytes = JsonEncodedText.Encode(KeySizePropertyName);
 
+        /// <summary>
+        /// The size of the RSA key, the value must be a valid RSA key length such as 2048 or 4092
+        /// </summary>
         public int? KeySize { get; set; }
 
+        /// <summary>
+        /// Create RsaKeyOptions specifying whether the certificate key should be stored in the HSM
+        /// </summary>
+        /// <param name="hsm">Specifies whether the certificate key should be stored in the HSM</param>
         public RsaKeyOptions(bool hsm)
             : base(hsm ? CertificateKeyType.RsaHsm : CertificateKeyType.Rsa)
         {
